Add files via upload
Git_0_1_Solution_1.ipynb - This file contains a solution to the base file (Git_0_1_copy.ipynb or the master file). In order to make it work you have to initialize all the steps before running the solution part. This file is created as a continue to the main file that we had to finish. Please note without activating the logical chain and all steps the solution part will not work properly.