Add files via upload
Git_0_1_Solutition_only.ipynb - This file is pure solution to the problem that we had to improve. Works without any additional operations. Its a solution by itself and works on its own. The code was re-written and adjusted, so everything is working as a single unit, does not contain any additional code parts or steps that prior should be activated. For review purpose only. The main part of this file was already integrated in Git_0_1_Solutition_1.ipynb which can be automatically merged with the master file after review.